Added more SETLL tests
